[file-manager] feat: add middleware to get auth info from header and insert into request object
