+++ file-manager/package.json
@@ -23,7 +23,6 @@
     "@types/cors": "2.8.17",
     "@types/express": "4.17.21",
     "@types/jest": "29.5.8",
-    "@types/jsonwebtoken": "^9.0.6",
     "@types/luxon": "3.3.7",
     "@types/multer": "^1.4.11",
     "@types/node": "20.9.0",
@@ -48,7 +47,6 @@
     "cors": "2.8.5",
     "dotenv": "16.3.1",
     "express": "4.18.2",
-    "jsonwebtoken": "^9.0.2",
     "kafkajs": "2.2.4",
     "luxon": "3.4.4",
     "multer": "^1.4.5-lts.1",
