+++ file-manager/src/presentation/rest/middleware/index.ts
@@ -1,0 +1,1 @@
+export * from './get-auth-info-from-headers'
