+++ file-manager/src/shared/env.ts
@@ -5,8 +5,7 @@ dotenv.config()
 export const ENVS = {
   APP: {
     PORT: process.env.PORT ?? 3001,
-    ID: process.env.APP_ID ?? 'file-manager-service',
-    TOKEN_SECRET: process.env.TOKEN_SECRET ?? ''
+    ID: process.env.APP_ID ?? 'file-manager-service'
   },
   KAFKA: {
     BROKERS_URL: process.env.KAFKA_HOSTS ? String(process.env.KAFKA_HOSTS).split(',') : [],
